+++ src/renderers/TypeScript.ts
@@ -12,6 +12,7 @@ import EnumDefinition from '../definitions/Enum';
 import InstanceDefinition from '../definitions/Instance';
 import NumberDefinition from '../definitions/Number';
 import ObjectDefinition from '../definitions/Object';
+import ReferenceDefinition from '../definitions/Reference';
 import ShapeDefinition from '../definitions/Shape';
 import StringDefinition from '../definitions/String';
 import UnionDefinition from '../definitions/Union';
@@ -19,6 +20,7 @@ import indent from '../helpers/indent';
 import isPrimitive from '../helpers/isPrimitive';
 import normalizeType from '../helpers/normalizeType';
 import { Config, Options, PrimitiveType } from '../types';
+import DefinitionFactory from '../DefinitionFactory';
 
 const ASCII_ALPHA_START: number = 65;
 
@@ -32,24 +34,17 @@ export default class TypeScriptRenderer extends Renderer {
   }
 
   renderArray(definition: ArrayDefinition, depth: number): string {
-    if (!definition.valueType) {
-      return 'any[]';
-    }
-
-    const configType = definition.valueType.config.type;
-    let template = this.renderAttribute(definition.valueType, depth);
-
-    if (isPrimitive(configType) || configType === 'instance') {
-      template += '[]';
-    } else {
-      template = this.wrapGenerics('Array', template);
-    }
-
-    return template;
+    return this.wrapNullable(
+      definition,
+      this.wrapGenerics(
+        'Array',
+        definition.valueType ? this.renderAttribute(definition.valueType, depth) : 'any',
+      ),
+    );
   }
 
   renderBool(definition: BoolDefinition): string {
-    return 'boolean';
+    return this.wrapNullable(definition, 'boolean');
   }
 
   renderEnum(definition: EnumDefinition, depth: number): string {
@@ -57,16 +52,14 @@ export default class TypeScriptRenderer extends Renderer {
     const { values, valueType } = definition.config;
     const members: string[] = [];
     const enumName = this.getObjectName(this.schematic.name, definition.attribute, 'Enum');
-    let currentIndex = 0;
     let currentChar = ASCII_ALPHA_START;
 
     switch (normalizeType(valueType)) {
-      // If a string is provided, assign values incrementally starting from 0
+      // If a string is provided, let TS infer the value (which is numeric)
       default:
       case 'string':
         values.forEach((value: PrimitiveType) => {
           members.push(`${indent(1, char)}${String(value)}`);
-          currentIndex += 1;
         });
         break;
 
@@ -82,45 +75,66 @@ export default class TypeScriptRenderer extends Renderer {
 
     this.builder.header.add(`export enum ${enumName} ${this.formatObject(members, 0, ',\n')}`);
 
-    return enumName;
+    return this.wrapNullable(definition, enumName);
   }
 
   renderInstance(definition: InstanceDefinition): string {
-    return this.formatValue(definition.config.contract, 'function');
+    return this.wrapNullable(definition, this.formatValue(definition.config.contract, 'function'));
   }
 
   renderNumber(definition: NumberDefinition): string {
-    return 'number';
+    return this.wrapNullable(definition, 'number');
   }
 
   renderObject(definition: ObjectDefinition, depth: number): string {
     const key = definition.keyType ? this.renderAttribute(definition.keyType, depth) : 'string';
     const value = definition.valueType ? this.renderAttribute(definition.valueType, depth) : 'any';
 
-    return this.formatObject(`[key: ${key}]: ${value}`, 0, ' ', ' ');
+    return this.wrapNullable(definition, this.formatObject(`[key: ${key}]: ${value}`, 0, ' ', ' '));
+  }
+
+  renderReference(definition: ReferenceDefinition): string {
+    return this.wrapNullable(definition, super.renderReference(definition));
   }
 
   renderShape(definition: ShapeDefinition, depth: number): string {
     const reference = this.renderShapeReference(definition);
 
     if (reference) {
-      return reference;
+      return this.wrapNullable(definition, reference);
     }
 
     if (!definition.attributes) {
-      return 'object';
+      return this.wrapNullable(definition, 'object');
     }
 
-    return this.formatObject(this.renderObjectProps(definition.attributes, depth + 1, ';'), depth);
+    return this.wrapNullable(
+      definition,
+      this.formatObject(this.renderObjectProps(definition.attributes, depth + 1, ';'), depth),
+    );
   }
 
   renderString(definition: StringDefinition): string {
-    return 'string';
+    return this.wrapNullable(definition, 'string');
   }
 
   renderUnion(definition: UnionDefinition, depth: number): string {
-    const set = new Set(definition.valueTypes.map(item => this.renderAttribute(item, depth)));
+    return this.wrapNullable(
+      definition,
+      Array.from(
+        new Set(definition.valueTypes.map(item => this.renderAttribute(item, depth))),
+      ).join(' | '),
+    );
+  }
 
-    return Array.from(set.values()).join(' | ');
+  /**
+   * Render a definition and wrap for TypeScript nullable support.
+   */
+  wrapNullable(definition: Definition<Config>, template: string): string {
+    if (definition.isNullable()) {
+      return `${template} | null`;
+    }
+
+    return template;
   }
 }