+++ tests/expected/gql/flow/union.js
@@ -11,12 +11,12 @@ export type UnionBarStructType = {
 };
 
 export type UnionBazStructType = {
-  baz: ?boolean | ?number | null,
+  baz: boolean | number | null,
 };
 
 export type UnionType = {
-  primitiveField: ?boolean | ?number | null,
+  primitiveField: boolean | number | null,
   enumField: 'FOO' | 'BAR' | 'BAZ' | 'QUX' | null,
-  shapeField: ?UnionFooStructType | ?UnionBarStructType | ?UnionBazStructType | null,
-  unionField: ?boolean | ?number | null | 'FOO' | 'BAR' | 'BAZ' | 'QUX' | null | ?UnionFooStructType | ?UnionBarStructType | ?UnionBazStructType | null,
+  shapeField: UnionFooStructType | UnionBarStructType | UnionBazStructType | null,
+  unionField: boolean | number | 'FOO' | 'BAR' | 'BAZ' | 'QUX' | UnionFooStructType | UnionBarStructType | UnionBazStructType | null,
 };